feat: refactor navigation imports and add new components for improved structure and functionality
