+++ app/[locale]/(home)/style.css
@@ -3,9 +3,7 @@
 }
 
 .bg-gradient {
-  background: #000000; /* fallback for old browsers */
-  background: -webkit-linear-gradient(45deg, #000000 70%, #434343); /* Chrome 10-25, Safari 5.1-6 */
-  background: linear-gradient(45deg, #000000 70%, #434343); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
+    @apply bg-gradient-to-b from-black to-gray-900;
 }
 
 .rainbow-border {
