+++ app/[locale]/(main)/counter.tsx
