+++ app/[locale]/(main)/spot-light.tsx
